Upgrade svelte to latest version
refs https://github.com/sveltejs/kit/discussions/5774

- This upgrade requires a migration to a new format/layout
