+++ package.json
@@ -2,35 +2,36 @@
   "name": "geojs-app",
   "version": "0.0.1",
   "scripts": {
-    "dev": "svelte-kit dev",
-    "build": "svelte-kit build",
+    "dev": "vite dev",
+    "build": "vite build",
     "package": "svelte-kit package",
-    "preview": "svelte-kit preview",
-    "check": "svelte-check --tsconfig ./tsconfig.json",
-    "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
-    "lint": "prettier --ignore-path .gitignore --check --plugin-search-dir=. . && eslint --ignore-path .gitignore .",
-    "format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. ."
+    "preview": "vite preview",
+    "check": "svelte-check --tsconfig ./jsconfig.json",
+    "check:watch": "svelte-check --tsconfig ./jsconfig.json --watch",
+    "lint": "prettier --check --plugin-search-dir=. . && eslint .",
+    "format": "prettier --write --plugin-search-dir=. ."
   },
   "devDependencies": {
-    "eslint": "8.17.0",
+    "eslint": "8.22.0",
     "eslint-config-prettier": "8.5.0",
     "eslint-plugin-svelte3": "4.0.0",
-    "prettier": "2.6.2",
-    "prettier-plugin-tailwindcss": "0.1.11"
+    "prettier": "2.7.1",
+    "prettier-plugin-tailwindcss": "0.1.13"
   },
   "type": "module",
   "dependencies": {
-    "@felte/reporter-svelte": "1.1.2",
-    "@sveltejs/adapter-auto": "1.0.0-next.50",
-    "@sveltejs/adapter-netlify": "1.0.0-next.64",
-    "@sveltejs/kit": "1.0.0-next.350",
+    "@felte/reporter-svelte": "1.1.3",
+    "@sveltejs/adapter-auto": "1.0.0-next.64",
+    "@sveltejs/adapter-netlify": "1.0.0-next.71",
+    "@sveltejs/kit": "1.0.0-next.409",
     "@tailwindcss/forms": "0.5.2",
-    "autoprefixer": "10.4.7",
-    "felte": "1.2.2",
-    "postcss": "8.4.14",
-    "svelte": "3.48.0",
-    "svelte-check": "2.7.1",
-    "svelte-preprocess": "4.10.6",
-    "tailwindcss": "3.0.24"
+    "autoprefixer": "10.4.8",
+    "felte": "1.2.3",
+    "postcss": "8.4.16",
+    "svelte": "3.49.0",
+    "svelte-check": "2.8.0",
+    "svelte-preprocess": "4.10.7",
+    "tailwindcss": "3.1.8",
+    "vite": "^3.0.0"
   }
 }
