+++ src/routes/s/+error.svelte
@@ -1,4 +1,4 @@
-<script context="module">
+<script>
 	import Search from '$lib/components/Search.svelte';
 </script>
